integrating login page to the existing frontend
Login page and routing moved to the existing frontend.
FrontEnd folder and its content was removed.
