--- Login Page -  Front End/Public/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Login/Sign Up</title>
-    <link rel="stylesheet" href="bootstrap.min.css">
-    <link rel="stylesheet" href="LoginPage.css" />
-  </head>
-  <body>
-    <div id="root"></div>
-    <script src="LoginPage.js"></script>
-  </body>
-</html>
