--- Login Page -  Front End/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "my-login-page",
-  "version": "1.0.0",
-  "description": "A simple login page built with React",
-  "main": "LoginPage.js",
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "login": "node ./src/LoginPage.js"
-  },
-  "keywords": [
-    "login",
-    "page",
-    "React"
-  ],
-  "author": "Your Name",
-  "license": "MIT",
-  "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
-}
