--- Login Page -  Front End/src/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import LoginPage from './LoginPage';
-
-function App() {
-  return (
-    <div className="App">
-      <LoginPage />
-    </div>
-  );
-}
-
-export default App;
