--- Login Page -  Front End/src/router.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Login from './components/Login';
-import Signup from './components/Signup';
-
-function AppRouter() {
-  return (
-    <Router>
-      <Switch>
-        <Route exact path="/" component={Login} />
-        <Route path="/signup" component={Signup} />
-      </Switch>
-    </Router>
-  );
-}
-
-export default AppRouter;
