+++ dashboard-frontend/package.json
@@ -14,6 +14,7 @@
     "bootstrap": "^5.3.0-alpha1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "recharts": "^2.5.0",
     "styled-components": "^5.3.9",
