+++ dashboard-frontend/src/App.js
@@ -1,9 +1,18 @@
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Dashboard from './components/Dashboard';
+import LoginPage from './components/LoginPage';
 
 function App() {
   return (
     <>
-      <Dashboard/>
+      <BrowserRouter>
+        <Routes>
+          <Route exact path="/" element={<LoginPage/>} />
+          {/* <Route path="/signup" element={<Signup/>} /> */}
+          <Route path="/dashboard" element={<Dashboard/>} />
+        </Routes>
+    </BrowserRouter>
     </>
   );
 }
