+++ dashboard-frontend/src/css/LoginPage.css
@@ -1,0 +1,126 @@
+.container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #1b1b6a;
+}
+
+.login-box {
+  position: relative;
+  width: 400px;
+  height: 350px;
+  padding: 40px;
+  background-color: #ffffff;
+  justify-content: center;
+  border-radius: 30px;
+}
+
+.login-box h2 {
+  margin: 0;
+  padding: 0 0 20px;
+  text-align: center;
+  font-size: 22px;
+  color: white;
+}
+
+.user-box {
+  position: relative;
+}
+
+.user-box input {
+  width: 100%;
+  padding: 10px 0;
+  font-size: 16px;
+  color: #333;
+  margin-bottom: 30px;
+  border: none;
+  border-bottom: 1px solid #ccc;
+  outline: none;
+  background: transparent;
+}
+
+.user-box label {
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding: 10px 0;
+  font-size: 16px;
+  color: #333;
+  pointer-events: none;
+  transition: 0.5s;
+}
+
+.user-box input:focus ~ label,
+.user-box input:valid ~ label {
+  top: -20px;
+  left: 0;
+  color: #03a9f4;
+  font-size: 12px;
+}
+
+.line {
+  width: 50%;
+  height: 1px;
+  background-color: #ccc;
+  border-bottom: 3px solid #1b1b6a;
+  margin: 2px auto;
+}
+
+.login-box a {
+  text-decoration: none;
+  font-size: 14px;
+  color: #333;
+  margin-left: 5px;
+}
+
+.login-box a:hover {
+  color: #03a9f4;
+}
+
+input[type="submit"] {
+  border: none;
+  outline: none;
+  height: 40px;
+  background: #03a9f4;
+  color: white;
+  font-size: 18px;
+  border-radius: 20px;
+  cursor: pointer;
+  transition: 0.5s;
+  margin-top: 40px;
+}
+
+input[type="submit"]:hover {
+  background: #0162c8;
+}
+
+input[type="submit"]:focus {
+  outline: none;
+}
+
+.signup {
+  color: white;
+  margin-bottom: 10px auto;
+}
+
+.signup a {
+  color: white;
+  text-decoration: underline;
+}
+
+.login-btn {
+  background-color: white;
+  color: navy;
+  border: none;
+  padding: 10px 20px;
+  font-size: 16px;
+  cursor: pointer;
+  transition: 0.3s;
+  margin-top: 20px auto;
+}
+
+.login-btn:hover {
+  background-color: #ddd;
+}
+
